Chisel3: signals declared correct cast not needed

+++ hardware/src/main/scala/patmos/Decode.scala
@@ -165,7 +165,7 @@ class Decode() extends Module {
   val isStack = Wire(Bool())
 
   val isSTC = Wire(Bool())
-  val stcImm = Cat(UInt(0), instr(17, 0), Bits("b00")).toUInt()
+  val stcImm = Cat(UInt(0), instr(17, 0), Bits("b00"))
 
   // Long immediates set this
   longImm := Bool(false)
@@ -357,7 +357,7 @@ class Decode() extends Module {
   // we could mux the imm / register here as well
 
   // Immediate for absolute calls
-  io.decex.callAddr := Cat(UInt(0), instr(21, 0), Bits("b00")).toUInt()
+  io.decex.callAddr := Cat(UInt(0), instr(21, 0), Bits("b00"))
 
   // Immediate for branch is sign extended, not extended for call
   // PC-relative value is precomputed here
